refactor: migrate blockSettings:editForm
